requirements.txt
srctag[embedding]==0.5.0
loguru
chromadb==0.4.14
PyGithub==2.1.1
tabulate==0.9.0